Resizing PartitionId and LogId and types
This PR reduces the size of PartitionId to a conservative u16 and LogId to u32. This attempts to fix a category of problems we often see when working with those generously-sized types. We would like to have tighter control on how many one of the types would fit into the other, for instance, we can share a bifrost log-id space with multiple clusters. Additionally, with a small partition-id space, we always have the option to expand it later, but with the current u64, it'd be very difficult to go the other way around.

This PR comes with a small risk of incompatibility if anyone has was able to run with more than 65k partitions on a single node. I don't think that's feasible with the amount of resources that it'd require, so, I deem this transition as (almost) safe. We would be performing further compatibility checks prior to release, but according to my local tests, I've not observed any issues.

+++ crates/admin/src/cluster_controller/scheduler.rs
@@ -900,7 +900,7 @@ mod tests {
 
     fn random_cluster_state(
         node_ids: &Vec<GenerationalNodeId>,
-        num_partitions: u64,
+        num_partitions: u16,
     ) -> ClusterState {
         let nodes = random_nodes_state(node_ids, num_partitions);
 
@@ -915,7 +915,7 @@ mod tests {
 
     fn random_nodes_state(
         node_ids: &Vec<GenerationalNodeId>,
-        num_partitions: u64,
+        num_partitions: u16,
     ) -> BTreeMap<PlainNodeId, NodeState> {
         let mut result = BTreeMap::default();
         let mut rng = rand::thread_rng();
@@ -949,7 +949,7 @@ mod tests {
     fn random_alive_node(
         rng: &mut ThreadRng,
         node_id: GenerationalNodeId,
-        num_partitions: u64,
+        num_partitions: u16,
     ) -> AliveNode {
         let partitions = random_partition_status(rng, num_partitions);
         AliveNode {
@@ -961,7 +961,7 @@ mod tests {
 
     fn random_partition_status(
         rng: &mut ThreadRng,
-        num_partitions: u64,
+        num_partitions: u16,
     ) -> BTreeMap<PartitionId, PartitionProcessorStatus> {
         let mut result = BTreeMap::default();
